Add files via upload
idn_east and idn_west are rectangles that split the region down the middle vertically. idn_mask_east and idn_mask_west are the resulting masks produced from an intersection of the original reef mask and the east/west rectangles. These geojsons are necessary to work within the constraints of our current mapping workflow code.